requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.50.0
    # via matplotlib
importlib-resources==6.4.0
    # via matplotlib
kiwisolver==1.4.5
    # via matplotlib
matplotlib==3.8.3
    # via -r requirements.in
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
packaging==24.0
    # via matplotlib
pillow==10.2.0
    # via
    #   -r requirements.in
    #   matplotlib
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
six==1.16.0
    # via python-dateutil
zipp==3.18.1
    # via importlib-resources
